Redact possible secrets in report
