+++ tests/scripts/generate-test-reports.ts
@@ -15,6 +15,7 @@ import * as fs from "fs";
 import * as path from "path";
 import { fileURLToPath } from "url";
 import { useAgentRunner, type AgentRunConfig } from "../utils/agent-runner";
+import { redactSecrets } from "../utils/redact";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -64,7 +65,7 @@ function filterSubdirectoriesBySkill(subdirectories: string[], skill: string): s
     // See tests/eslint-rules/integration-test-name.mjs for details.
     const terminatorIndex = subdirName.indexOf("_");
     const skillName = subdirName.substring(0, terminatorIndex);
-    
+
     return skillName === skill;
   });
 }
@@ -117,6 +118,8 @@ async function processSubdirectory(subdirPath: string, reportTemplate: string): 
 
     consolidatedContent += `\n## ${fileName}\n\n${content}\n`;
   }
+
+  consolidatedContent = redactSecrets(consolidatedContent);
 
   console.log("    Generating report...");
